# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bonus App API",
        "description": "REST API documentation for - https://ziyoforum-app.uz",
        "contact": {
            "email": "uktamovich95@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://ziyoforum-app.uz/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User login",
                "description": "User login",
                "operationId": "userLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserToken"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequest"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User logout",
                "description": "User logout",
                "operationId": "userLogout",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get list of news",
                "description": "Returns list of news",
                "operationId": "getNewsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/news/{id}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get news by id",
                "description": "Return news by id",
                "operationId": "getNewsById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "News id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user by token",
                "description": "Get user by token",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BadRequest": {
                "title": "Bad Request",
                "description": "Bad Request",
                "properties": {
                    "error": {
                        "title": "Error message",
                        "description": "Error message",
                        "type": "string",
                        "example": "Error text"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Bad Request"
                }
            },
            "News": {
                "title": "News",
                "description": "News model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "Title",
                        "description": "News title",
                        "type": "string",
                        "example": "A nice news"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the new project",
                        "type": "string",
                        "example": "This is new project's description"
                    },
                    "content": {
                        "title": "Content",
                        "description": "Content of the new project",
                        "type": "string",
                        "example": "This is new project's Content"
                    },
                    "image_url": {
                        "title": "Image path",
                        "description": "Image path",
                        "type": "string",
                        "example": "https://ziyoforum-app.uz/storage/19/360x160.png"
                    },
                    "published_at": {
                        "title": "Published at",
                        "description": "Published at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "News"
                }
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "first_name": {
                        "title": "First name",
                        "description": "First name",
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "title": "Last name",
                        "description": "Last name",
                        "type": "string",
                        "example": "Smith"
                    },
                    "patronymic": {
                        "title": "Patronymic",
                        "description": "Patronymic",
                        "type": "string",
                        "example": "David"
                    },
                    "phone": {
                        "title": "Phone",
                        "description": "Phone",
                        "type": "string",
                        "example": "998939604040"
                    },
                    "gender": {
                        "title": "Gender",
                        "description": "Gender",
                        "type": "string",
                        "example": "male"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email",
                        "type": "string",
                        "example": "example@gmail.com"
                    },
                    "email_verified": {
                        "title": "Email verify",
                        "description": "Email verify",
                        "type": "boolean",
                        "example": "true"
                    },
                    "role": {
                        "$ref": "#/components/schemas/UserRole"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
            "UserRole": {
                "title": "User role",
                "description": "User role",
                "properties": {
                    "role_id": {
                        "title": "Role id",
                        "description": "Role id",
                        "type": "integer",
                        "format": "int64",
                        "example": 3
                    },
                    "role_name": {
                        "title": "Role name",
                        "description": "Role name",
                        "type": "string",
                        "example": "user"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User role"
                }
            },
            "UserToken": {
                "title": "User token",
                "description": "User token",
                "properties": {
                    "token": {
                        "title": "Token",
                        "description": "User token",
                        "type": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9ib251cy1hcHAudGVzdFwvYXBpXC92MVwvdXNlclwvbG9naW4iLCJpYXQiOjE2NjE0NTk0MTEsImV4cCI6MTY2MTQ2MzAxMSwibmJmIjoxNjYxNDU5NDExLCJqdGkiOiJCdzBkRUFZeUFvZzVXQXk1Iiwic3ViIjo3LCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.ekZBR8qs1OAdU_uQAkV4DAkwdfetaaYc4HbJj0ogO8U"
                    },
                    "expire_in": {
                        "title": "Token expire",
                        "description": "oken expire",
                        "type": "integer",
                        "format": "int64",
                        "example": 60
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User token"
                }
            },
            "NewsCollection": {
                "title": "NewsCollection",
                "description": "News collection",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/News"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "NewsCollection"
                }
            },
            "NewsResource": {
                "title": "NewsResource",
                "description": "News resource",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/News"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "NewsResource"
                }
            },
            "UserResource": {
                "title": "UserResource",
                "description": "User resource",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserResource"
                }
            },
            "UserLoginRequest": {
                "title": "User login request",
                "description": "User login request body data",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "Email",
                        "type": "string",
                        "example": "example@mail.ru"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password",
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object"
            }
        }
    }
}